refactor: split blog-helpers and add page libs
