+++ src/components/header.tsx
@@ -4,7 +4,7 @@ import Head from 'next/head'
 import classnames from 'classnames'
 import ExtLink from './ext-link'
 import { meta, navigation as navItems } from '../../site.config'
-import { getCurrentPageTitle, isActivePage } from '../libs/client/blog-helpers'
+import { getCurrentPageTitle, isActivePage } from '../libs/util'
 import faviconIcon from '../../public/favicon.ico'
 
 const themeClassMap = {
