+++ src/pages/api/sitemap.ts
@@ -7,7 +7,7 @@ import { Readable } from 'stream'
 import pMap from 'p-map'
 import * as dayjs from 'dayjs'
 import utc from 'dayjs/plugin/utc'
-import { getSinglePagePath } from '../../libs/client/blog-helpers'
+import { getSinglePagePath } from '../../libs/notion'
 import log from '../../libs/server/log'
 import { getNotionPostsFromTable } from '../../libs/server/notion'
 import { navigation, notion, cache as cacheConfig } from '../../../site.config'
